Rhythm
player can no receive extra damage if the press a designated button within the timeFrame of the chosen key-moment of a song.
For example if a key-moment in a song is at 01:11 (71 seconds) and that keyMoment plays in the game-world and the player presses the SpaceBar for example 3 seconds within the key-moment, they will receive a boost to their attack. Three seconds after the designated key-moment the damage goes back to normal